Explicitly set column types for non-empty tables as well as empty tables
- insert_all() infers column types based on data, but when a column has no data (nulls), the default is TEXT.
- this is an issue for columns that are meant to be INTEGER foreign keys.

Approach: examine the source table's schema and set the column types (just like the empty table situation)

Note: the schema query is done before the select query to avoid interfering with the rows being retrieved in the generator.

(Doing it later caused problems when the source database was SQL Server).